# Conflicts:
#	.github/workflows/ci.yml
#	composer.lock
#	legacy/classes/cs_mail.php
#	legacy/classes/cs_material_item.php
#	legacy/functions/date_functions.php
#	src/Controller/ContextController.php
#	src/Controller/GroupController.php
#	src/Controller/ItemController.php
#	src/Controller/MaterialController.php
#	src/Controller/PortalSettingsController.php
#	src/Controller/TodoController.php
#	src/Controller/UserController.php
#	src/Cron/Tasks/CronExpireTakeOver.php
#	src/EventSubscriber/AutoRoomMembershipSubscriber.php
#	src/EventSubscriber/LoginSubscriber.php
#	src/EventSubscriber/SecuritySubscriber.php
#	src/EventSubscriber/UserLocaleSubscriber.php
#	src/Form/DataTransformer/GeneralSettingsTransformer.php
#	src/Form/Type/Profile/AccountContactFormType.php
#	src/Form/Type/SendListType.php
#	src/Mail/Mailer.php
#	src/Metrics/AccountLoginsMetric.php
#	src/Utils/MailAssistant.php
#	templates/base.html.twig
#	tests/Unit/AutoRoomMembershipSubscriberTest.php

+++ legacy/classes/cs_group_item.php
@@ -69,7 +69,7 @@ class cs_group_item extends cs_label_item
     /** save news item
      * this methode save the news item into the database.
      */
-    public function save($save_other = true)
+    public function save(bool $saveGrouproom = true): void
     {
         $current_user_item = null;
         $save_time = false;
@@ -78,7 +78,7 @@ class cs_group_item extends cs_label_item
         $parentRoom = $this->getContextItem();
         $portal = $parentRoom->getContextItem();
 
-        if ($save_other) {
+        if ($saveGrouproom) {
             if (!$this->_issetGroupRoomItemID() && $this->isGroupRoomActivated()) {
                 $new_group_room = true;
 
@@ -114,7 +114,7 @@ class cs_group_item extends cs_label_item
                 // Zeitpunkte
                 $save_time = $portal->showTime();
 
-                $grouproom_item->saveOnlyItem();
+                $grouproom_item->save(false);
 
                 // add member of group to the group room
                 $current_user_item = $this->_environment->getCurrentUserItem();
@@ -191,9 +191,9 @@ class cs_group_item extends cs_label_item
             $grouproom_item->setLogoFilename($disc_manager->getLastSavedFileName());
             $save2 = true;
         }
-        if (isset($save2) and $save2 and $save_other) {
+        if (isset($save2) and $save2 and $saveGrouproom) {
             $grouproom_item->setLinkedGroupItemID($this->getItemID());
-            $grouproom_item->saveOnlyItem();
+            $grouproom_item->save(false);
         }
         if (isset($new_group_room) and $new_group_room) {
             $this->setGroupRoomItemID($grouproom_item->getItemID());
@@ -208,24 +208,19 @@ class cs_group_item extends cs_label_item
         $this->updateElastic();
     }
 
-    /** save news item
-     * this methode save the news item into the database.
-     */
-    public function saveOnlyItem()
-    {
-        $this->save(false);
-    }
-
     /** delete group item
      * this methode delete the group item
      * with the group room.
      */
-    public function delete()
+    public function delete(bool $deleteGrouproom = true): void
     {
-        $room = $this->getGroupRoomItem();
-        if (isset($room)) {
-            $room->delete();
+        if ($deleteGrouproom) {
+            $room = $this->getGroupRoomItem();
+            if (isset($room)) {
+                $room->delete();
+            }
         }
+
         parent::delete();
     }
